Run launcher diagnostics in production Node runtime
